requirements.txt
discord.py
flask
python-dotenv
ollama
faiss-cpu
sentence-transformers
